miApiJWT/requirements.txt
fastapi[standard]
uvicorn[standard]
PyJWT
python-multipart